+++ crates/release_plz/tests/all/helpers/fake_utils.rs
@@ -1,0 +1,7 @@
+use fake::{StringFaker, Fake};
+
+pub fn fake_id() -> String {
+    const LETTERS: &str = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ";
+    let f = StringFaker::with(Vec::from(LETTERS), 8);
+    f.fake()
+}
