+++ crates/release_plz/tests/all/helpers/gitea/mod.rs
@@ -14,4 +14,5 @@ pub struct GiteaContext {
     pub token: String,
     /// Repository name.
     repo: String,
+    client: reqwest::Client,
 }
