+++ crates/release_plz/tests/all/helpers/mod.rs
@@ -1,3 +1,4 @@
-mod context;
+pub mod test_context;
 pub mod gitea;
 mod reqwest_utils;
+mod fake_utils;
